wip: remove unused files and fix submit form error
