+++ apps/withdraw-cow-amm/src/app/layout.tsx
@@ -1,18 +1,28 @@
 "use client";
 
-import { RootLayout } from "@bleu/cow-hooks-ui";
+import { useGetHookInfo } from "#/hooks/useGetHookInfo";
+import { RootLayout, WithdrawFormContextProvider } from "@bleu/cow-hooks-ui";
 import "@bleu/cow-hooks-ui/global.css";
 import Head from "next/head";
 
 import type * as React from "react";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
+  const getHookInfo = useGetHookInfo();
+
   return (
     <html lang="en" className="h-screen">
       <Head>
         <link rel="manifest" href="/manifest.json" />
       </Head>
-      <RootLayout>{children}</RootLayout>
+      <RootLayout>
+        <WithdrawFormContextProvider
+          getHookInfo={getHookInfo}
+          poolTypeIn="COW_AMM"
+        >
+          {children}
+        </WithdrawFormContextProvider>
+      </RootLayout>
     </html>
   );
 }
