+++ packages/cow-hooks-ui/src/hooks/useUserPools.ts
@@ -1,17 +1,23 @@
-import type { SupportedChainId } from "@cowprotocol/cow-sdk";
-
 import { BigNumber } from "ethers";
-import type { Address } from "viem";
 import { usePools } from "./usePools";
+import { useIFrameContext } from "../context/iframe";
 
-export function useUserPools(chainId?: SupportedChainId, user?: Address) {
+export function useUserPools(poolType: "WEIGHTED" | "COW_AMM") {
+  const { context } = useIFrameContext();
+  const protocolVersionIn: (1 | 2)[] = poolType === "WEIGHTED" ? [2] : [1];
   const useSwrData = usePools(
-    { poolTypeIn: ["COW_AMM"], userAddress: user },
-    chainId,
-    "userbalanceUsd",
+    {
+      poolTypeIn: [poolType],
+      userAddress: context?.account,
+      protocolVersionIn,
+    },
+    context?.chainId,
+    "userbalanceUsd"
   );
+
   const data = useSwrData.data?.filter((pool) =>
-    BigNumber.from(pool.userBalance.walletBalance).gt(BigNumber.from("10")),
+    BigNumber.from(pool.userBalance.walletBalance).gt(BigNumber.from("10"))
   );
+
   return { ...useSwrData, data };
 }
