--- packages/utils/jest.config.js
@@ -1,7 +1,0 @@
-/* eslint-env node */
-/** @type {import('ts-jest').JestConfigWithTsJest} */
-module.exports = {
-  preset: "ts-jest",
-  testEnvironment: "node",
-  workerThreads: true,
-};
